fix: refactored for experimental resource picker
